convenience-store-api/requirements.txt
asgiref==3.7.2
Django==3.2.23
django-cors-headers==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==9.5.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
